requirements.txt
fastapi==0.128.0
uvicorn==0.32.0
jose==1.0.0
sqlalchemy==2.0.45
asyncpg==0.31.0
cryptography==46.0.3
pymongo==4.16.0
dotenv==0.9.9
pytest==9.0.2
httpx==0.28.1
pytest-asyncio==1.3.0
pyjwt==2.10.1
pydantic_settings==2.12.0
bcrypt==5.0.0
aioboto3==15.5.0
pillow==12.1.0
python-multipart==0.0.22
python-box==7.3.2
pyyaml==6.0.3